+++ frontend/cloud/src/containers/organizations/userDetail.js
@@ -12,6 +12,7 @@ import { loadDomainUserRoles, saveUserRole, deleteUserRole } from "../../modules
 import { SUPERADMIN_RIGHTS_RESOURCE, ADMIN_RIGHTS_RESOURCE, loadUserRights } from "../../modules/validations"
 import { open } from "shared/modules/alert"
 import Spinner from "shared/containers/spinner"
+import { confirmationDialog } from "shared/utils"
 
 class UserDetail extends React.Component {
     constructor(props) {
@@ -103,15 +104,20 @@ class UserDetail extends React.Component {
         }
     }
 
-    deleteUserRole(userRoleID) {
+    deleteUserRole(index) {
         return e => {
-            // if there's no more organizationUserRoles after removal, remove user from organization
-            if (_.values(this.props.organizationUserRoles).length === 1) {
-                this.props.deleteUserFromOrganization(this.props.organizationID, this.props.userID)
-                this.props.history.push(`/organizations/${this.props.organizationID}`)
-            } else {
-                this.props.deleteUserRole(userRoleID)
-            }
+            confirmationDialog(
+                `Click OK to confirm that you want to remove role ${this.props.roles[this.state.userRoles[index].roleID].name} from the user.`,
+                () => {
+                    // if there's no more organizationUserRoles after removal, remove user from organization
+                    if (_.values(this.props.organizationUserRoles).length === 1) {
+                        this.props.deleteUserFromOrganization(this.props.organizationID, this.props.userID)
+                        this.props.history.push(`/organizations/${this.props.organizationID}`)
+                    } else {
+                        this.props.deleteUserRole(this.state.userRoles[index].id)
+                    }
+                }
+            )
         }
     }
 
@@ -194,7 +200,7 @@ class UserDetail extends React.Component {
                                             </div>
                                         ) : (
                                             <div className="btn-group" role="group">
-                                                <button className="btn btn-link" type="button" onClick={this.deleteUserRole(userRole.id)}>
+                                                <button className="btn btn-link" type="button" onClick={this.deleteUserRole(i)}>
                                                     <span className="remove-link">Remove</span>
                                                 </button>
                                             </div>
